Function wayland_scanner::generate_code

pub fn generate_code<P1: AsRef<Path>, P2: AsRef<Path>>(
    prot: P1, 
    target: P2, 
    side: Side
)

Generate the code for a protocol

See this crate toplevel documentation for details.

Args:

protocol: a path to the XML file describing the protocol, absolute or relative to the build script using this function.

target: the path of the file to store the code in.

side: the side (client or server) to generate code for.